Generative_AI_Project/requirements.txt
diffusers==0.19.0
transformers==4.30.0
torch==2.2.0
accelerate==0.22.0
matplotlib==3.7.0
Pillow==9.3.0
